requirements.txt
Flask
python-dotenv
Flask-SQLALchemy
flask_wtf
flask 
Werkzeug